requirements.txt
django==2.2.10
djangorestframework
django-rest-registration
coreapi
pygments
markdown